backend/requirements.txt
django
psycopg2
requests
python-decouple
django-environ
djangorestframework
drf-yasg 
pillow 
python-magic
cryptography